style: bulk ruff format + I001 import sort (192 files)
Applied automatically via 'ruff format src/cohezion/' and
'ruff check --fix --unsafe-fixes' (I001 + RUF013 passes).
No logic changes — pure formatting and import ordering.

Entire-Checkpoint: e3dcbfe6a642

+++ src/cohezion/arc/codec.py
@@ -97,7 +97,7 @@ def _denormalize_grid(arr: np.ndarray, rows: int, cols: int) -> Grid:
     for r in range(rows):
         row = []
         for c in range(cols):
-            val = int(round(arr[r, c] * 9.0))
+            val = round(arr[r, c] * 9.0)
             row.append(max(0, min(9, val)))
         out.append(row)
     return out
